+++ packages/react-jss/src/types.js
@@ -1,5 +1,5 @@
 // @flow
-import type {StyleSheetFactoryOptions, Jss, SheetsRegistry, SheetsManager} from 'jss'
+import type {StyleSheetFactoryOptions, Jss, SheetsRegistry, SheetsManager, BaseRule} from 'jss'
 import type {Node} from 'react'
 import type {Theming} from 'theming'
 
@@ -30,6 +30,10 @@ export type HOCProps<Theme, Props> = Props & {
 export type InnerProps = {
   children?: Node,
   classes: {}
+}
+
+export type DynamicRules = {
+  [key: string]: BaseRule
 }
 
 export type ThemedStyles<Theme> = (theme: Theme) => StaticStyles